-Refactoring code acording to Airbnb Design Guide. -Adding .editorconfig file to root folder. -Adding Rusty for running verification processes. Now it should run a lint check on commits, and a test check on pushes.
