+++ package.json
@@ -23,6 +23,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
+    "lint": "eslint --ext .jsx,.js src",
+    "pre-commit": "npm run lint",
+    "pre-push": "npm test",
+    "post-merge": "npm install", 
+    "post-rewrite": "npm install",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public"
   },
@@ -74,6 +79,7 @@
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-storybook": "^0.5.11",
+    "husky": "^8.0.1",
     "webpack": "^5.72.1"
   }
 }
