+++ src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import GlobalStyles from './global';
-import Header from './components/header';
-import Footer from './components/footer';
-import Content from './components/content';
+import { Header } from './components/header';
+import { Footer } from './components/footer';
+import { Content } from './components/content';
 
-const App = ():JSX.Element => {
+function App():React.ReactElement {
   return (
     <>
       <GlobalStyles />
