+++ src/components/content/formSection/index.ts
@@ -1,0 +1,2 @@
+export { default as FormSection } from './FormSection';
+export * from './styles';
