+++ src/components/content/formSection/styles.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const FormWrapper = styled.div`
+  display: auto;
+`;
