+++ src/components/content/fullTextSection/index.ts
@@ -1,0 +1,2 @@
+export { default as FullTextSection } from './FullTextSection';
+export * from './styles';
