+++ src/components/content/heroSection/index.ts
@@ -1,0 +1,2 @@
+export { default as HeroSection } from './HeroSection';
+export * from './styles';
