--- src/components/content/in/formSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from './FormSection';
-export * from './styles';
