--- src/components/content/in/formSection/styles.tsx
@@ -1,2 +1,0 @@
-import styled from "styled-components";
-
