--- src/components/content/in/fullTextSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from './FullTextSection';
-export * from './styles';
