--- src/components/content/in/heroSection/index.ts
@@ -1,2 +1,0 @@
-export { default } from './HeroSection';
-export * from './styles';
