+++ src/components/footer/index.ts
@@ -1,2 +1,2 @@
-export { default } from './Footer';
+export { default as Footer } from './Footer';
 export * from './styles';
