+++ src/components/header/Header.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { HeaderWrapper } from '.';
 import { Button, Logo } from '../common/stylesMixing';
-import NavList from './in/navList';
+import { NavList } from './navList';
 
-const Header = ():JSX.Element => {
-
+function Header():React.ReactElement {
   return (
     <HeaderWrapper>
       <div className="logo-wrapper">
@@ -15,7 +14,7 @@ const Header = ():JSX.Element => {
         <Button>CURRÍCULUM</Button>
       </div>
     </HeaderWrapper>
-  )
+  );
 }
 
 export default Header;
