--- src/components/header/in/navList/index.ts
@@ -1,2 +1,0 @@
-export { default } from './NavList';
-export * from './styles';
