+++ src/components/header/index.ts
@@ -1,2 +1,2 @@
-export { default } from './Header';
+export { default as Header } from './Header';
 export * from './styles';
