+++ src/components/header/navItem/index.ts
@@ -1,0 +1,2 @@
+export { default as NavItem } from './NavItem';
+export * from './styles';
