+++ src/components/header/navList/index.ts
@@ -1,0 +1,2 @@
+export { default as NavList } from './NavList';
+export * from './styles';
