+++ tsconfig.json
@@ -22,5 +22,6 @@
   },
   "include": [
     "src"
-  ]
+  ],
+  "baseUrl": "src",
 }
